+++ regenerate.sh
@@ -14,10 +14,10 @@ tv_scripts=(
     orchard_poseidon
     orchard_poseidon_hash
     orchard_sinsemilla
-    orchardzsa_asset_base
-    orchardzsa_issuance_auth_sig
-    orchardzsa_key_components
-    orchardzsa_note_encryption
+    orchard_zsa_asset_base
+    orchard_zsa_issuance_auth_sig
+    orchard_zsa_key_components
+    orchard_zsa_note_encryption
     sapling_generators
     sapling_key_components
     sapling_note_encryption
