+++ zcash_test_vectors/orchard_zsa/issuance_auth_sig.py
@@ -2,7 +2,7 @@
 import sys;
 
 from zcash_test_vectors.bip340_reference import pubkey_gen, schnorr_sign, hash_sha256
-from zcash_test_vectors.orchardzsa.key_components import IssuanceKeys
+from zcash_test_vectors.orchard_zsa.key_components import IssuanceKeys
 
 from ..output import render_args, render_tv
 
@@ -52,7 +52,7 @@ def main():
 
     render_tv(
         args,
-        'orchardzsa_issuance_auth_sig',
+        'orchard_zsa_issuance_auth_sig',
         (
             ('isk', '[u8; 32]'),
             ('ik', '[u8; 32]'),
